Overhaul the download process: resolve, then retrieve. Change APIs.
Split the download process up into "Resolve" and "Retrieve":
  - "Resolve" fetches information for the URLs and enqueue the resulting `File` objects for Downloading.
  - "Retrieve" downloads those files in parallel, as usual, but using the obtained information.
Downloader:
  - Work endlessly, waiting for jobs from queue.
  - Remove `StartSync` and `StartAsync`.
  - Return a WaitGroup for provided links instead which the caller can wait for.
DownloadQueue:
  - Wait for instructions in a busy loop and work endlessly.
  - Drop external channel dependencies.
  - Use standard container for the internal queue which will enable manipulating downloads in a future version.
  - Move `AddLinks` into Downloader and don't expose Queue.
Move the job handler for accounts to a shared file and reuse it in DownloadQueue.
Copy Go's `Once` type and tailor it to the project's requirements.
Remove FileSpec and replace it with the new File interface.
Remove Action as it is not needed anymore.
Remove Bundle as it is not used (might return in the future).