Put error in File. Fix priorities for "Bundle" requests.
File: Put errors in core.File interface to facilitate better handling.
Priorities: don't use math.Nextafter. The use of math.Nextafter turned out to be very flawed for grandchildren requests. Particularly, if request #0 bundles 10 requests #1-10, #3 of which bundles another 2 requests #11-12, this request #12 will have the same priority as request #4, which of course is not desired. Instead, an equidistant gap should be found between the two sibling requests, however, this requires a lot of work and tracking. The priorities work on a more sophisticated level now, recursing up the ancestry to determine the preceding node.
CLI: Use File#Err() to determine a file error. Additionally, make it clear in the summary if there was a file that has unknown length.
Add documentation.